+++ README.md
@@ -1,11 +1,14 @@
 # xM API SDK JS
+
 `xmas` for short 🎄
 
 # Usage
-If your project already relies on the `axios` library,
-`xmas` will just use it to send http requests to `xmApi` and handle its responses.
+
+If your project already relies on the `axios` library, `xmas` will just use it to send http requests
+to `xmApi` and handle its responses.
 
 Your instantiation of a new Xmas object will only need your xM hostname and some auth credentials:
+
 ```js
 const Xmas = require('xmas');
 
@@ -24,8 +27,9 @@ xmas.groups.create(group)
   .catch(handleError);
 ```
 
-Alternative `config` object for **OAuth**
-(say when you have already generated tokens and safely stored them in your DB):
+Alternative `config` object for **OAuth** (say when you have already generated tokens and safely
+stored them in your DB):
+
 ```json
 {
   "hostname": "https://yourOrg.xmatters.com",
@@ -36,6 +40,7 @@ Alternative `config` object for **OAuth**
 ```
 
 ## Obtain OAuth tokens
+
 ```js
 const Xmas = require('xmas');
 
@@ -51,12 +56,15 @@ xmas.getOauthTokens.byUsernamePassword()
   .then(() => xmas.people.search('immediately uses the tokens, not the creds set in config'))
   .catch(handleError);
 ```
-`Xmas` will immediately start using the OAuth tokens and stop using the username & password
-you instantiated it with.
+
+`Xmas` will immediately start using the OAuth tokens and stop using the username & password you
+instantiated it with.
 
 ## Dependency injection
-If your project relies on an **HTTP client** *other* than `axios`,
-you will need to pass it in the `config` when you instantiate an Xmas:
+
+If your project relies on an **HTTP client** _other_ than `axios`, you will need to pass it in the
+`config` when you instantiate an Xmas:
+
 ```js
 const config = {
   hostname: 'https://yourOrg.xmatters.com',
@@ -66,46 +74,55 @@ const config = {
     sendRequest: yourHttpClient,
     successAdapter: () => {},
     failureAdapter: () => {},
-  }
+  },
 };
 ```
 
 ### httpClient.sendRequet
+
 Should have the following signature:
+
 ```js
-({ method, url, headers, data }) => Promise
+(({ method, url, headers, data }) => Promise);
 ```
+
 Where:
-+ `method` will be an HTTP method used to send the request (eg: 'GET', 'POST', 'DELETE')
-+ `url` will be the fully qualified url the request will be sent to
-(eg: 'https://yourOrg.xmatters.com/api/xm/1/people?firstName=peter&lastName=parker')
-+ `headers` will be a typical HTTP request headers object to send to xM API
-+ `data` (optional) will be the stringified payload to send to xM API
+
+- `method` will be an HTTP method used to send the request (eg: 'GET', 'POST', 'DELETE')
+- `url` will be the fully qualified url the request will be sent to (eg:
+  'https://yourOrg.xmatters.com/api/xm/1/people?firstName=peter&lastName=parker')
+- `headers` will be a typical HTTP request headers object to send to xM API
+- `data` (optional) will be the stringified payload to send to xM API
 
 Your HTTP client should know what to do with those and must return a `promise`.
 
 ### httpClient.successAdapter
-This is a function that will receive the response
-in the exact format your HTTP client usually returns it upon a successful request (2xx).
+
+This is a function that will receive the response in the exact format your HTTP client usually
+returns it upon a successful request (2xx).
 
 Think the very first `.then()` called when your HTTP client promise `resolves`.
 
-This function must *only* return the xmApi `payload`/`response body`.
+This function must _only_ return the xmApi `payload`/`response body`.
 
 Here is an example of the adapter used for the axios HTTP client under the hood:
+
 ```js
 const axiosSuccessAdapter = (res) => res.data;
 ```
 
 ### httpClient.failureAdapter
-This is a function that will receive the error
-in the exact format your HTTP client usually throws it upon a failed request (non 2xx).
+
+This is a function that will receive the error in the exact format your HTTP client usually throws
+it upon a failed request (non 2xx).
 
 Think the `.catch()` called when your HTTP client promise `rejects`.
 
-This function must throw (rethrow, technically) an error object with both a `status` and a `payload` property attached to it.
+This function must throw (rethrow, technically) an error object with both a `status` and a `payload`
+property attached to it.
 
 Here is an example of the adapter used for the axios HTTP client under the hood:
+
 ```js
 const axiosFailureAdapter = (e) => {
   const humanReadableMessage = e.response
@@ -117,13 +134,13 @@ const axiosFailureAdapter = (e) => {
   throw error;
 };
 ```
+
 The human readable message can be omitted and the object thrown doesn't even have to be an Error
-instance, these are just nice things.
-What is important is that this function `throws`,
-and that the object thrown contains a `status` and a `payload` property.
-Where:
-+ `status` must be an *integer*: the http **status code** of the response
-+ `payload` must be the xM API **response body** if one was returned
+instance, these are just nice things. What is important is that this function `throws`, and that the
+object thrown contains a `status` and a `payload` property. Where:
+
+- `status` must be an _integer_: the http **status code** of the response
+- `payload` must be the xM API **response body** if one was returned
 
 ```sh
 # If all of this seems like more trouble than having to manage 1 more dependency in your project,
